Hero Section Setup Added
