+++ frontend/src/Components/Pages/Landing-Page/HeroSection.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+// Design Hero section here and call this hero section in "frontend/src/Pages/Landing-Page.jsx"
+
+const HeroSection = () => {
+	return (
+		<div className="w-full h-full text-cyan-50 text-2xl text-center">
+			Hero Section
+		</div>
+	);
+};
+
+export default HeroSection;
