--- frontend/src/Components/Pages/Landing-Page/LandingHeader.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-const Header = () => {
-  return (
-    <div className="w-full h-full text-cyan-50 text-2xl text-center">
-      Header Section
-    </div>
-  );
-};
-
-export default Header;
