+++ frontend/src/Pages/Landing-Page.jsx
@@ -1,17 +1,17 @@
 import React from "react";
-import LandingHeader from "../Components/Pages/Landing-Page/LandingHeader";
-import Feature from "../Components/Pages/Landing-Page/Feature";
+
 import MeetingCards from "../Components/Pages/Landing-Page/MeetingCards";
 import OurTeam from "../Components/Pages/Landing-Page/OurTeam";
 import JoinRoom from "./JoinRoom-Page";
+import HeroSection from "../Components/Pages/Landing-Page/HeroSection";
 
 export const LandingPage = () => {
 	return (
 		<div>
-			<LandingHeader />
+			<HeroSection />
 			<JoinRoom />
 			<MeetingCards />
-			<Feature />
+			<HeroSection />
 			<OurTeam />
 		</div>
 	);
